+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `-s` argument to package genstructs command
+- `to_one` gadget that merges all SUI coins 'to one' for an address
 
 ### Fixed